requirements.txt
albucore==0.0.20
albumentations==1.4.21
altgraph==0.17.4
annotated-types==0.7.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dnspython==2.7.0
einops==0.8.0
entmax==1.3
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.0
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.55.0
fsspec==2024.10.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
latex2mathml==3.77.0
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
MouseInfo==0.1.3
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pefile==2023.2.7
pillow==11.0.0
pix2tex==0.1.2
prov==2.0.1
puremagic==1.28
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pymongo==4.10.1
PyMsgBox==1.0.9
PyMuPDF==1.24.13
pyparsing==3.2.0
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
simplejson==3.19.3
simsimd==6.0.5
six==1.16.0
stringzilla==3.10.10
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.5.4
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traits==6.4.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
x-transformers==0.15.0
